Comments and printing to screen
Several comments included which helped me work out how the program
worked.
After I worked out how to do it, I then went about printing the contents
of the two constructed vectors to the screen.

At this point there's an error - names ends up with 1 more entry than
score and so when I ran a loop based on names.size() and came to
retrieving score, there was an out of range assert. Need to fix that.